blueprints/legacy/remote-terminal-python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "remote-terminal-agent"
version = "1.0.0"
description = "Remote terminal agent using Python"
requires-python = ">=3.10"
dependencies = [
    "websockets>=12.0",
    "msgpack>=1.0.0",
    "pywinpty>=2.0.0;platform_system=='Windows'",
    "ptyprocess>=0.7.0;platform_system!='Windows'",
    "httpx>=0.27.0",
]

[project.scripts]
remote-terminal-agent = "remote_terminal.main:main"

[tool.hatch.build.targets.wheel]
packages = ["remote_terminal"]
